requirements.txt
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
idna==2.10
isodate==0.6.0
lxml==4.6.3
psycopg2-binary==2.9.3
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.2
python-dotenv==0.19.0
rdflib==5.0.0
requests==2.25.1
schedule==1.0.0
selenium==3.141.0
six==1.16.0
smbprotocol==1.9.0
urllib3==1.26.5
